updated file structure & css cleaned-up
