--- src/components/About/about.css
@@ -1,4 +1,0 @@
-.about {
-  margin: 0;
-  padding: 0 2rem ;
-}
