--- src/components/Adverts/Adverts.js
@@ -1,13 +1,0 @@
-import React from "react";
-import "./otomoto.css";
-
-const Adverts = () => {
-  return (
-    <div className="otomoto">
-      <h2>Nasze ogłoszenia na OtoMoto.pl</h2>
-      <p>Strona aktualnie w budowie.</p>
-    </div>
-  );
-};
-
-export default Adverts;
