--- src/components/Adverts/otomoto.css
@@ -1,4 +1,0 @@
-.otomoto {
-  /* margin: 50px; */
-  padding: 0 2rem;
-}
