+++ src/components/InfoTile/InfoTile.js
@@ -1,9 +1,10 @@
 import React from "react";
+import "./infoTile.css";
 
 const InfoTile = (props) => {
   return (
-    <div className="serviceTile">
-      <img src={props.icon}></img>
+    <div className="infoTile">
+      <img src={props.icon} alt="props.desc"></img>
       <h3>{props.desc}</h3>
     </div>
   );
