--- src/components/Main/main.css
@@ -1,49 +1,0 @@
-.mainTitle {
-  font-family: "Alfa Slab One", cursive;
-  text-transform: uppercase;
-  font-size: 4rem;
-  color: var(--gray);
-  text-align: center;
-}
-
-.subTitle {
-  font-family: "Caveat", cursive;
-  font-size: 3rem;
-  color: var(--gray);
-  text-align: center;
-}
-
-section {
-  padding: 0.5rem 0;
-  /* box-shadow: var(--titleShadow); */
-}
-
-.services {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-}
-
-.serviceTile {
-  align-content: center;
-  overflow: hidden;
-  flex: 1;
-  height: 20vw;
-  padding: 0px;
-  margin: 1rem 0;
-  text-align: center;
-  /* box-shadow: 0 0 1px 1px var(--gray); */
-}
-
-.serviceTile img {
-  align-self: center;
-  height: 80%;
-  margin: 0 0;
-
-  /* height: 100%; */
-}
-
-.serviceTile h3 {
-  height: 20%;
-  margin: 0 0;
-}
