--- src/components/Map/map.css
@@ -1,21 +1,0 @@
-.map {
-  margin: 0;
-  padding: 2rem;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  height: 75vh;
-}
-
-.map-frame {
-  margin: 1rem 0;
-  position: relative;
-  width: 100%;
-  height: 100%;
-  /* height: 80%; */
-  padding-bottom: 0;
-  margin-bottom: 0;
-  border-radius: var(--corners);
-  box-shadow: var(--shadow);
-}
