+++ src/index.css
@@ -49,3 +49,9 @@ h2 {
   color: var(--gray);
   box-shadow: var(--titleShadow);
 }
+
+.infoTilesBox {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+}
