+++ src/pages/Adverts/Adverts.js
@@ -1,0 +1,15 @@
+import React from "react";
+import "./otomoto.css";
+
+const Adverts = () => {
+  return (
+    <div className="container">
+      <section>
+        <h2>Nasze ogłoszenia na OtoMoto.pl</h2>
+        <p>Strona aktualnie w budowie.</p>
+      </section>
+    </div>
+  );
+};
+
+export default Adverts;
