+++ src/pages/Main/main.css
@@ -1,0 +1,18 @@
+.mainTitle {
+  font-family: "Alfa Slab One", cursive;
+  text-transform: uppercase;
+  font-size: 4rem;
+  color: var(--gray);
+  text-align: center;
+}
+
+.subTitle {
+  font-family: "Caveat", cursive;
+  font-size: 3rem;
+  color: var(--gray);
+  text-align: center;
+}
+
+section {
+  padding: 0.5rem 0;
+}
