+++ src/pages/Map/index.js
@@ -1,2 +1,3 @@
 import Map from "./Map";
 export default Map;
+//test
